+++ Source/AbyssDiverUnderWorld/AnimNodes/AnimNode_AquaticBoneTrail.cpp
@@ -4,7 +4,7 @@
 #include "AnimNodes/AnimNode_AquaticBoneTrail.h"
 #include "Animation/AnimInstanceProxy.h"
 #include "Components/SkeletalMeshComponent.h"
-#include "Monster/AquaticMovementComponent.h"
+#include "Monster/Components/AquaticMovementComponent.h"
 #include "GameFramework/Character.h"
 #include "DrawDebugHelpers.h"
