+++ Source/AbyssDiverUnderWorld/Monster/EQS/EQG_XYZGrid.cpp
@@ -7,7 +7,7 @@
 #include "NavigationSystem.h"
 #include "EnvironmentQuery/Items/EnvQueryItemType_Point.h"
 #include "Monster/Boss/Boss.h"
-#include "Monster/AquaticMovementComponent.h"
+#include "Monster/Components/AquaticMovementComponent.h"
 
 
 UEQG_XYZGrid::UEQG_XYZGrid()
