+++ Source/AbyssDiverUnderWorld/Monster/Monster.cpp
@@ -12,7 +12,7 @@
 #include "Character/UnderwaterCharacter.h"
 #include "Components/CapsuleComponent.h"
 #include "Interactable/OtherActors/Radars/RadarReturn2DComponent.h"
-#include "Monster/AquaticMovementComponent.h"
+#include "Monster/Components/AquaticMovementComponent.h"
 #include "Monster/Components/TickControlComponent.h"
 
 const FName AMonster::MonsterStateKey = "MonsterState";
